Rename hook js file to match kebab case of hook name Prevent trying to edit an actor in a locked compendium
