+++ module/hooks/index.js
@@ -1,6 +1,7 @@
-import * as Init from './init.js'
+import * as CreateToken from './create-token.js'
 import * as DiceSoNiceReady from './dice-so-nice-ready.js'
 import * as DiceSoNiceRollStart from './dice-so-nice-roll-start.js'
+import * as Init from './init.js'
 import * as Ready from './ready.js'
 import * as RenderActorSheet from './render-actor-sheet.js'
 import * as RenderChatMessage from './render-chat-message.js'
@@ -9,10 +10,12 @@ import * as RenderItemSheet from './render-item-sheet.js'
 import * as RenderPause from './render-pause.js'
 import * as RenderPlayerList from './render-player-list.js'
 import * as RenderSettingsConfig from './render-settings-config.js'
-import * as TokenCreation from './token-creation.js'
 
 export const CoC7Hooks = {
   listen () {
+    CreateToken.listen()
+    DiceSoNiceReady.listen()
+    DiceSoNiceRollStart.listen()
     Init.listen()
     Ready.listen()
     RenderActorSheet.listen()
@@ -20,10 +23,7 @@ export const CoC7Hooks = {
     RenderDialog.listen()
     RenderItemSheet.listen()
     RenderPause.listen()
-    DiceSoNiceReady.listen()
-    DiceSoNiceRollStart.listen()
     RenderPlayerList.listen()
     RenderSettingsConfig.listen()
-    TokenCreation.listen()
   }
 }
